requirements.txt
Flask
flask-cors
openai
python-dotenv
PyPDF2
numpy
